Modif, début suppr apres date de remise

+++ src/index.js
@@ -1,33 +1,43 @@
 const Discord = require('discord.js');
 const botClient = new Discord.Client();
 const config = require('../config.json');
-const asciiCats = require('ascii-cats')
 
-let db = require('./devoirs.json');
+const asciiCats = require('ascii-cats')
 
 const embed = require('./embeds');
 const ajout = require('./ajoutDB')
 const utils = require('./utils')
 const init = require('./initDb')
 const suppr = require('./supprDB')
+const modif = require('./modifDB')
 
+let db = require('./devoirs.json');
+
+/**
+ * Au démarrage du bot
+ */
 botClient.on('ready', () => {
+	//Status du bot
 	botClient.user.setActivity("!help-agenda");
-	console.clear();
 
+	console.clear();
 	console.log("=============================")
 	console.log(asciiCats('nyan'));
 	console.log("\n\n Bot started !")
 	console.log("=============================")
 });
 
+/**
+ * Des qu'un message sur le serveur ou le bot est présent est reçu
+ */
 botClient.on('message', msg => {
+	suppr.supprDevoirDate(db, msg);
 
+	//On regarde si le message commence bien par le prefix (!)
 	if (!msg.content.startsWith(config.prefix))//Si le message ne commence pas par le prefix du config.json
 		return;
 
-	switch (msg.content.substr(1).split(" ")[0]) {
-
+	switch (msg.content.substr(1).split(" ")[0]) {//Switch sur le premier mot du msg sans le prefix Ex: "!agenda dejfez" donne "agenda"
 		case 'init-agenda':
 			init.groupInit(db, msg);
 			break;
@@ -37,15 +47,11 @@ botClient.on('message', msg => {
 			break;
 
 		case 'debug':
-			msg.reply("```" + JSON.stringify(db, null, 4) + "```")
-			msg.delete();
+			utils.debugDbFile(db, msg);
 			break;
 
 		case 'clear-db':
-			msg.delete();
-			db = { "groups": [] };
-			utils.updateDbFile(db);
-			console.warn("DATABASE RESET");
+			db = utils.clearDbFile(db, msg)
 			break;
 
 		case 'help-agenda':
@@ -56,20 +62,22 @@ botClient.on('message', msg => {
 			suppr.supprDb(db, msg);
 			break;
 
+		case 'modif-agenda' :
+			modif.modifDB(db, msg);
+			break;
+
 		default:
-			console.log("commande non reconnue");
+			console.log("Commande non reconnue : " + msg.content);
 			break;
 	}
 
-	try {
-		msg.delete()
-	}
-	catch {
-		(e) => {
-			console.error(e);
-		}
-	}
+	if (msg)
+		if (msg.deletable)
+			msg.delete()
+				.then(e => { console.log("Message supprimé") })
+				.catch(e => { console.error("Message déjà supprimé") });
 
 });
 
+//Bot login
 botClient.login(config.token);
